+++ Uebung2/lab2/src/main/java/at/ac/tuwien/big/we16/ue2/service/BiddingServlet.java
@@ -41,7 +41,7 @@ public class BiddingServlet extends HttpServlet {
                     .findAny();
 
             //set new balance
-            float balanceUpdated = user.getBalance() - newPrice;
+            double balanceUpdated = user.getBalance() - newPrice;
             if (existingBid.isPresent()) {
                 balanceUpdated += existingBid.get().getPrice();
             }
